RT#61144 Appropriately handle non-object jid
This applies the patch submitted in the aformentioned ticket, to
appropriately handle instances $jid (et al.) is a string rather
than an object. In particular, this avoids Perl vomit when the jid
starts with a number. The original roster.t test was modified to
break before application of the patch. It now passes.